--- a/llms/mlx_lm/models/mixtral.py
+++ b/llms/mlx_lm/models/mixtral.py
@@ -1,9 +1,9 @@
+import math
 from dataclasses import dataclass
 from typing import Dict, Optional, Tuple, Union
 
 import mlx.core as mx
 import mlx.nn as nn
-import numpy as np
 
 from .base import BaseModelArgs
 
@@ -91,22 +91,129 @@ class MixtralAttention(nn.Module):
         return self.o_proj(output)
 
 
-class MixtralBLockSparseTop2MLP(nn.Module):
-    def __init__(self, args: ModelArgs):
+class QuantizedSwitchMLP(nn.Module):
+    def __init__(self, dim, hidden_dim, num_experts, group_size=64, bits=4):
         super().__init__()
-        self.ffn_dim = args.intermediate_size
-        self.hidden_dim = args.hidden_size
+        scale_in = math.sqrt(1 / dim)
+        scale_hidden = math.sqrt(1 / hidden_dim)
+        self.gate_proj_w, self.gate_proj_s, self.gate_proj_b = mx.quantize(
+            mx.random.uniform(
+                low=-scale_in,
+                high=scale_in,
+                shape=(num_experts, hidden_dim, dim),
+            ),
+            group_size=group_size,
+            bits=bits,
+        )
+        self.up_proj_w, self.up_proj_s, self.up_proj_b = mx.quantize(
+            mx.random.uniform(
+                low=-scale_in,
+                high=scale_in,
+                shape=(num_experts, hidden_dim, dim),
+            ),
+            group_size=group_size,
+            bits=bits,
+        )
+        self.down_proj_w, self.down_proj_s, self.down_proj_b = mx.quantize(
+            mx.random.uniform(
+                low=-scale_hidden,
+                high=scale_hidden,
+                shape=(num_experts, dim, hidden_dim),
+            ),
+            group_size=group_size,
+            bits=bits,
+        )
+        self.num_experts = num_experts
+        self.group_size = group_size
+        self.bits = bits
 
-        self.w1 = nn.Linear(self.hidden_dim, self.ffn_dim, bias=False)
-        self.w2 = nn.Linear(self.ffn_dim, self.hidden_dim, bias=False)
-        self.w3 = nn.Linear(self.hidden_dim, self.ffn_dim, bias=False)
+    def __call__(self, x, indices) -> mx.array:
+        x = mx.expand_dims(x, (-2, -3))
+        x3 = mx.block_sparse_qmm(
+            x,
+            self["w3_w"],
+            self["w3_s"],
+            self["w3_b"],
+            rhs_indices=indices,
+            transpose=True,
+            group_size=self.group_size,
+            bits=self.bits,
+        )
+        x1 = mx.block_sparse_qmm(
+            x,
+            self["w1_w"],
+            self["w1_s"],
+            self["w1_b"],
+            rhs_indices=indices,
+            transpose=True,
+            group_size=self.group_size,
+            bits=self.bits,
+        )
+        y = mx.block_sparse_qmm(
+            nn.silu(x1) * x3,
+            self["w2_w"],
+            self["w2_s"],
+            self["w2_b"],
+            rhs_indices=indices,
+            transpose=True,
+            group_size=self.group_size,
+            bits=self.bits,
+        )
+        return y.squeeze(-2)
 
-        self.act_fn = nn.silu
 
-    def __call__(self, x: mx.array) -> mx.array:
-        current_hidden_states = self.act_fn(self.w1(x)) * self.w3(x)
-        current_hidden_states = self.w2(current_hidden_states)
-        return current_hidden_states
+class SwitchMLP(nn.Module):
+    def __init__(self, dim, hidden_dim, num_experts):
+        super().__init__()
+        scale_in = math.sqrt(1 / dim)
+        scale_hidden = math.sqrt(1 / hidden_dim)
+        self.w1 = mx.random.uniform(
+            low=-scale_in,
+            high=scale_in,
+            shape=(num_experts, hidden_dim, dim),
+        )
+        self.w2 = mx.random.uniform(
+            low=-scale_hidden,
+            high=scale_hidden,
+            shape=(num_experts, dim, hidden_dim),
+        )
+        self.w3 = mx.random.uniform(
+            low=-scale_in,
+            high=scale_in,
+            shape=(num_experts, hidden_dim, dim),
+        )
+        self.num_experts = num_experts
+
+    def __call__(self, x, indices) -> mx.array:
+        x = mx.expand_dims(x, (-2, -3))
+        x3 = mx.block_sparse_mm(x, self.w3.swapaxes(-1, -2), rhs_indices=indices)
+        x1 = mx.block_sparse_mm(
+            x,
+            self.w1.swapaxes(-1, -2),
+            rhs_indices=indices,
+        )
+        return mx.block_sparse_mm(
+            nn.silu(x1) * x3,
+            self.w2.swapaxes(-1, -2),
+            rhs_indices=indices,
+        ).squeeze(-2)
+
+    def to_quantized(self, group_size=64, bits=4):
+        num_experts, hidden_dim, dim = self.w1.shape
+        qm = QuantizedSwitchMLP(dim, hidden_dim, num_experts)
+        qm.w1_w, qm.w1_s, qm.w1_b = mx.quantize(
+            self.w1, group_size=group_size, bits=bits
+        )
+        qm.w2_w, qm.w2_s, qm.w2_b = mx.quantize(
+            self.w2, group_size=group_size, bits=bits
+        )
+        qm.w3_w, qm.w3_s, qm.w3_b = mx.quantize(
+            self.w3, group_size=group_size, bits=bits
+        )
+        return qm
+
+    def is_quantized(self, weights, prefix):
+        return f"{prefix}.w1_s" in weights
 
 
 class MixtralSparseMoeBlock(nn.Module):
@@ -120,43 +227,20 @@ class MixtralSparseMoeBlock(nn.Module):
         # gating
         self.gate = nn.Linear(self.hidden_dim, self.num_experts, bias=False)
 
-        self.experts = [
-            MixtralBLockSparseTop2MLP(args=args) for _ in range(self.num_experts)
-        ]
+        self.switch_mlp = SwitchMLP(self.hidden_dim, self.ffn_dim, self.num_experts)
 
     def __call__(self, x: mx.array) -> mx.array:
-        ne = self.num_experts_per_tok
-        orig_shape = x.shape
-        x = x.reshape(-1, x.shape[-1])
-
         gates = self.gate(x)
 
-        inds = mx.stop_gradient(mx.argpartition(-gates, kth=ne - 1, axis=-1)[:, :ne])
+        k = self.num_experts_per_tok
+        inds = mx.stop_gradient(mx.argpartition(-gates, kth=k - 1, axis=-1)[..., :k])
+        scores = mx.take_along_axis(gates, inds, axis=-1)
+        scores = mx.softmax(scores, axis=-1, precise=True)
 
-        scores = mx.softmax(
-            mx.take_along_axis(gates, inds, axis=-1).astype(mx.float32),
-            axis=-1,
-        ).astype(gates.dtype)
+        y = self.switch_mlp(x, inds)
+        y = (y * scores[..., None]).sum(axis=-2)
 
-        if self.training:
-            inds = np.array(inds)
-            y = mx.zeros((x.shape[0], ne, x.shape[-1]), x.dtype)
-            for e, expert in enumerate(self.experts):
-                idx1, idx2 = map(mx.array, np.where(inds == e))
-                if idx1.size == 0:
-                    continue
-                y[idx1, idx2] = expert(x[idx1])
-
-            y = (y * scores[:, :, None]).sum(axis=1)
-        else:
-            y = []
-            for xt, st, it in zip(x, scores, inds.tolist()):
-                yt = mx.stack([self.experts[e](xt) for e in it], axis=-1)
-                yt = (yt * st).sum(axis=-1)
-                y.append(yt[None, :])
-            y = mx.concatenate(y)
-
-        return y.reshape(orig_shape)
+        return y
 
 
 class MixtralDecoderLayer(nn.Module):
@@ -234,6 +318,19 @@ class Model(nn.Module):
     ):
         out = self.model(inputs, cache)
         return self.lm_head(out)
+
+    def sanitize(self, weights):
+        if "model.layers.0.block_sparse_moe.experts.0.w1.weight" not in weights:
+            return weights
+        for l in range(self.args.num_hidden_layers):
+            prefix = f"model.layers.{l}"
+            for n in ["w1", "w2", "w3"]:
+                to_join = [
+                    weights.pop(f"{prefix}.block_sparse_moe.experts.{e}.{n}.weight")
+                    for e in range(self.args.num_local_experts)
+                ]
+                weights[f"{prefix}.block_sparse_moe.switch_mlp.{n}"] = mx.stack(to_join)
+        return weights
 
     @property
     def layers(self):