@angeloskath 2024-05-20
@awni I think this is ready for a review. I didn't implement the `DoRASwitchLinear` but it can easily be added in a later PR that also adds quantization to DoRA layers.